requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
crispy-bootstrap5==2024.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0
django-allauth==64.2.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-environ==0.11.2
django-import-export==4.1.1
django-jazzmin==3.0.0
et-xmlfile==1.1.0
idna==3.9
openpyxl==3.1.5
psycopg2-binary==2.9.9
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
